Rename internal rand.h file
Replacement fix for #3975
