+++ crypto/include/internal/rand_int.h
